+++ src/test/kotlin/net/pwall/json/schema/codegen/CodeGeneratorArrayTest.kt
@@ -73,8 +73,7 @@ data class TestArray(
 ) {
 
     init {
-        require(aaa.isNotEmpty()) { "aaa length < minimum 1 - ${'$'}{aaa.size}" }
-        require(aaa.size <= 5) { "aaa length > maximum 5 - ${'$'}{aaa.size}" }
+        require(aaa.size in 1..5) { "aaa length not in range 1..5 - ${'$'}{aaa.size}" }
     }
 
     data class Person(
@@ -119,10 +118,8 @@ public class TestArray {
     ) {
         if (aaa == null)
             throw new IllegalArgumentException("Must not be null - aaa");
-        if (aaa.size() < 1)
-            throw new IllegalArgumentException("aaa length < minimum 1 - " + aaa.size());
-        if (aaa.size() > 5)
-            throw new IllegalArgumentException("aaa length > maximum 5 - " + aaa.size());
+        if (aaa.size() < 1 || aaa.size() > 5)
+            throw new IllegalArgumentException("aaa length not in range 1..5 - " + aaa.size());
         this.aaa = aaa;
         if (bbb == null)
             throw new IllegalArgumentException("Must not be null - bbb");